FATN_to_USB/pyproject.toml
[tool.poetry]
name = "fatn_to_usb"
version = "0.1.0"
description = "Download FATN zip file and load onto USB stick"
authors = ["Alistair Amabile <alistairanywhere@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
argparse = "^1.4.0"
Pillow = "^8.0.1"
requests = "^2.25.0"
ST7789 = "^0.0.2"
ST7735 = "^0.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
